Datenstruktur

+++ app/src/main/java/de/syntax_institut/lernkarten/team/ui/data/SpacedRepetitionData.kt
@@ -4,6 +4,6 @@ package de.syntax_institut.lernkarten.team.ui.data
 data class SpacedRepetitionData(
     val flashcardId: Long,
     var easeFactor: Double = 2.5, // Startwert für den Leichtigkeitsfaktor
-    var intervalDays: Int = 0,    // Tage bis zur nächsten Wiederholung
+    var intervalDays: Double = 0.0,    // Tage bis zur nächsten Wiederholung
     var consecutiveCorrect: Int = 0 // Anzahl der aufeinanderfolgenden korrekten Antworten
 )
